tinyllama1b/python_proj/requirements.txt
fastapi
llama_cpp-python